requirements.txt
mediapipe==0.9.1.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
pillow==10.2.0
streamlit==1.32.2
